Added web_scraper.py file for CodeClause Internship
This is a simple Python mini project I created for my CodeClause Internship.

It collects the latest news headlines from the NDTV website by using the requests and BeautifulSoup libraries. After running the script, it neatly displays the top 10 headlines directly in the terminal.

The project shows how Python can be used for basic web scraping tasks in a clean and efficient way.